Codereview

+++ IGRPhotoTweaks/IGRCropCornerView.swift
@@ -11,7 +11,7 @@ import UIKit
 class IGRCropCornerView: UIView {
 
     init(cornerType type: CropCornerType) {
-        super.init(frame: CGRect(x: CGFloat(0), y: CGFloat(0), width: CGFloat(kCropViewCornerLength), height: CGFloat(kCropViewCornerLength)))
+        super.init(frame: CGRect(x: 0.0, y: 0.0, width: kCropViewCornerLength, height: kCropViewCornerLength))
         
         setup(cornerType: type)
     }
@@ -25,28 +25,30 @@ class IGRCropCornerView: UIView {
     fileprivate func setup(cornerType type: CropCornerType) {
         self.backgroundColor = UIColor.clear
         let lineWidth: CGFloat = 2
-        let horizontal = UIView(frame: CGRect(x: CGFloat(0), y: CGFloat(0), width: CGFloat(kCropViewCornerLength), height: lineWidth))
+        
+        let horizontal = UIView(frame: CGRect(x: 0.0, y: 0.0, width: kCropViewCornerLength, height: lineWidth))
         horizontal.backgroundColor = UIColor.cropLine()
         self.addSubview(horizontal)
-        let vertical = UIView(frame: CGRect(x: CGFloat(0), y: CGFloat(0), width: lineWidth, height: CGFloat(kCropViewCornerLength)))
+        
+        let vertical = UIView(frame: CGRect(x: 0.0, y: 0.0, width: lineWidth, height: kCropViewCornerLength))
         vertical.backgroundColor = UIColor.cropLine()
         self.addSubview(vertical)
         
         if type == .upperLeft {
-            horizontal.center = CGPoint(x: CGFloat(kCropViewCornerLength / 2), y: CGFloat(lineWidth / 2))
-            vertical.center = CGPoint(x: CGFloat(lineWidth / 2), y: CGFloat(kCropViewCornerLength / 2))
+            horizontal.center = CGPoint(x: (kCropViewCornerLength / 2.0), y: (lineWidth / 2.0))
+            vertical.center = CGPoint(x: (lineWidth / 2.0), y: (kCropViewCornerLength / 2.0))
         }
         else if type == .upperRight {
-            horizontal.center = CGPoint(x: CGFloat(kCropViewCornerLength / 2), y: CGFloat(lineWidth / 2))
-            vertical.center = CGPoint(x: CGFloat(kCropViewCornerLength - lineWidth / 2), y: CGFloat(kCropViewCornerLength / 2))
+            horizontal.center = CGPoint(x: (kCropViewCornerLength / 2.0), y: (lineWidth / 2.0))
+            vertical.center = CGPoint(x: (kCropViewCornerLength - lineWidth / 2.0), y: (kCropViewCornerLength / 2.0))
         }
         else if type == .lowerRight {
-            horizontal.center = CGPoint(x: CGFloat(kCropViewCornerLength / 2), y: CGFloat(kCropViewCornerLength - lineWidth / 2))
-            vertical.center = CGPoint(x: CGFloat(kCropViewCornerLength - lineWidth / 2), y: CGFloat(kCropViewCornerLength / 2))
+            horizontal.center = CGPoint(x: (kCropViewCornerLength / 2.0), y: (kCropViewCornerLength - lineWidth / 2.0))
+            vertical.center = CGPoint(x: (kCropViewCornerLength - lineWidth / 2.0), y: (kCropViewCornerLength / 2.0))
         }
         else if type == .lowerLeft {
-            horizontal.center = CGPoint(x: CGFloat(kCropViewCornerLength / 2), y: CGFloat(kCropViewCornerLength - lineWidth / 2))
-            vertical.center = CGPoint(x: CGFloat(lineWidth / 2), y: CGFloat(kCropViewCornerLength / 2))
+            horizontal.center = CGPoint(x: (kCropViewCornerLength / 2.0), y: (kCropViewCornerLength - lineWidth / 2.0))
+            vertical.center = CGPoint(x: (lineWidth / 2.0), y: (kCropViewCornerLength / 2.0))
         }
     }
 }
